sai-arts-api/requirements.txt
djangorestframework
psycopg2
Pillow
pytz 
django-probes